Add commands under topic 'inspect' & rewrite log command
