+++ src/commands/aem/rde/inspect/request-logs/disable.js
@@ -15,7 +15,7 @@ const {
   BaseCommand,
   cli,
   commonFlags,
-} = require('../../../../lib/base-command');
+} = require('../../../../../lib/base-command');
 
 class DisableRequestLogsCommand extends BaseCommand {
   async run() {
