+++ src/commands/aem/rde/inspect/request-logs/index.js
@@ -11,37 +11,36 @@
  */
 'use strict';
 
-const { BaseCommand, cli, commonFlags } = require('../../../lib/base-command');
+const {
+  BaseCommand,
+  cli,
+  commonFlags,
+} = require('../../../../../lib/base-command');
 
-class SlingRequestsCommand extends BaseCommand {
+class RequestLogsCommand extends BaseCommand {
   async run() {
-    const { args, flags } = await this.parse(SlingRequestsCommand);
+    const { args, flags } = await this.parse(RequestLogsCommand);
     try {
       if (!args.id) {
         let params = {};
         params.filter = flags.include;
 
         let response = await this.withCloudSdk((cloudSdkAPI) =>
-          cloudSdkAPI.getSlingRequests(flags.target, params)
+          cloudSdkAPI.getRequestLogs(flags.target, params)
         );
         if (response.status === 200) {
           let json = await response.json();
-          cli.log('- Sling Requests: ');
-          json.items.forEach((slingRequests) => {
-            cli.log(slingRequests);
-          });
+          cli.log(JSON.stringify(json, null, 2));
         } else {
           cli.log(`Error: ${response.status} - ${response.statusText}`);
         }
       } else {
         let response = await this.withCloudSdk((cloudSdkAPI) =>
-          cloudSdkAPI.getSlingRequest(flags.target, args.id)
+          cloudSdkAPI.getRequestLog(flags.target, args.id)
         );
-
         if (response.status === 200) {
-          let slingRequest = await response.json();
-          cli.log(`- Sling Request "${args.id}": `);
-          cli.log(slingRequest);
+          let requestLog = await response.json();
+          cli.log(requestLog);
         } else {
           cli.log(`Error: ${response.status} - ${response.statusText}`);
         }
@@ -52,13 +51,13 @@ class SlingRequestsCommand extends BaseCommand {
   }
 }
 
-Object.assign(SlingRequestsCommand, {
+Object.assign(RequestLogsCommand, {
   description:
-    'Get the list of sling-requests for the target of a rapid development environment.',
+    'Get the list of request-logs for the target of a rapid development environment.',
   args: [
     {
       name: 'id',
-      description: 'The id of the sling-request to get.',
+      description: 'The id of the request-log to get.',
     },
   ],
   flags: {
@@ -67,4 +66,4 @@ Object.assign(SlingRequestsCommand, {
   },
 });
 
-module.exports = SlingRequestsCommand;
+module.exports = RequestLogsCommand;
